create fetch for displaying all current tournaments
